TradePulse/requirements.txt
aiohttp==3.8.6
alpha_vantage==2.3.1
feedparser==6.0.10
lightgbm==4.1.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.0
pandas==2.1.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.1
statsmodels==0.14.0
ta==0.10.2
transformers==4.34.0
yfinance==0.2.31
async-timeout==4.0.3
python-dateutil==2.8.2
praw==7.7.1
plotly==5.18.0
joblib==1.3.2
tqdm==4.66.1 